Adding serialization (Only tokens + scores for now).
