+++ tokenizers/src/models/unigram/mod.rs
@@ -1,5 +1,6 @@
 pub mod lattice;
 pub mod model;
+pub mod serialization;
 pub mod trainer;
 pub mod trie;
 pub mod unicode;
